requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.1
requests==2.31.0
pydantic==2.0.0
